feat: make modular structure
