+++ static/index.html
@@ -19,6 +19,8 @@
     <meta name="twitter:title" content="ListenBrainz Wrapped">
     <meta name="twitter:description" content="Open-source ListenBrainz Wrapped generator with Last.fm artwork and custom uploads.">
     <meta name="twitter:image" content="https://github.com/user-attachments/assets/ee64a7f1-6bbc-4af3-9a1b-3129de9c1f9c">
+    <link rel="icon" href="favicon.svg" type="image/svg+xml">
+    <link rel="alternate icon" href="img/pink.png" type="image/png">
     <script type="application/ld+json">
       {
         "@context": "https://schema.org",
@@ -35,7 +37,7 @@
       }
     </script>
     <link rel="stylesheet" href="style.css">
-    <script defer src="request.js"></script>
+    <script type="module" src="js/main.js"></script>
   </head>
   <body>
     <main class="page">
