+++ wrapped_fm/__init__.py
@@ -1,0 +1,5 @@
+"""Spotify Wrapped application package."""
+
+from .app import create_app
+
+__all__ = ["create_app"]
